+++ audiobooks/__init__.py
@@ -1,3 +1,3 @@
 """My audiobook library program."""
 
-__version__ = "v0.11.0"
+__version__ = "v0.11.1"